Physics and Engineering Actors

Physics and engineering actors are the fundamental building blocks of FUSE simulations:

Actors operate exclusively on IMAS.dd data

Actors functionality is controlled via act parameters

Actors can be combined into other actors

Fidelity hierarchy is enabled by concept of generic Vs specific actors

Generic actors define physics/component

Specific actors implement a specific model for that physics/component

For example:
ActorEquilibrium  <--  generic
├─ ActorSolovev   <--  specific
└─ ActorCHEASE    <--  specific

act.[ActorGeneric].model selects specific actor being used

All specific actors will expect data and fill the same enties in dd

IMAS.jl expressions are key to make this work seamlessly

Where possible actors should make use of generic actors and not hardcode use of specific actors

Physics and Engineering Actors

Balance plant (3 actors)

BalanceOfPlant

ThermalPlant

PowerNeeds

Build (5 actors)

CXbuild

HFSsizing

FluxSwing

LFSsizing

Stresses

Compound (3 actors)

DynamicPlasma

StationaryPlasma

WholeFacility

Control (1 actors)

ControllerIp

Costing (3 actors)

Costing

CostingARIES

CostingSheffield

Current (3 actors)

Current

QED

SteadyStateCurrent

Divertors (1 actors)

Divertors

Equilibrium (5 actors)

CHEASE

EGGO

Equilibrium

FRESCO

TEQUILA

Hcd (5 actors)

HCD

NeutralFueling

SimpleIC

SimpleLH

SimplePL

Nuclear (2 actors)

Blanket

Neutronics

Pedestal (3 actors)

Pedestal

EPED

WPED

Pf (3 actors)

PFdesign

PFactive

PassiveStructures

Stability (2 actors)

PlasmaLimits

VerticalStability

Transport (8 actors)

CoreTransport

FluxCalculator

FluxMatcher

AnalyticTurbulence

EPEDprofiles

Neoclassical

QLGYRO

TGLF

Wall loading (2 actors)

CoreRadHeatFlux

ParticleHeatFlux

Balance plant (3 actors)

BalanceOfPlant

FUSE.ActorBalanceOfPlant — Method

ActorBalanceOfPlant(dd::IMAS.dd, act::ParametersAllActors; kw...)

Balance of plant actor that estimates the net electrical power output by comparing the balance of plant electrical needs with the electricity generated from the thermal cycle.

Note

Stores data in dd.balance_of_plant

ActorBalanceOfPlant

ThermalPlant

FUSE.ActorThermalPlant — Method

ActorThermalPlant(dd::IMAS.dd, act::ParametersAllActors; kw...)

Note

Provides a common interface to run different thermal plant actors

ActorThermalPlant
├─ model
├─ fixed_plant_efficiency
├─ do_plot
└─ verbose

PowerNeeds

FUSE.ActorPowerNeeds — Method

ActorPowerNeeds(dd::IMAS.dd, act::ParametersAllActors; kw...)

Power needs actor that calculates the needed power to operate the plant

model = :thermal_power_fraction simply assumes that the power to balance a plant is a fraction of the gross electrical power generated by the thermal cycle.

model = :EU_DEMO subdivides the power plant electrical needs to [:cryostat, :tritium_handling, :pumping] using EU-DEMO numbers.

model = :FUSE subdivides power plant needs into subsystems and calculates their power needs.

Note

Stores data in dd.balance_of_plant.power_electric_plant_operation

ActorPowerNeeds
├─ model
├─ thermal_power_fraction
└─ do_plot

Build (5 actors)

CXbuild

FUSE.ActorCXbuild — Method

ActorCXbuild(dd::IMAS.dd, act::ParametersAllActors; kw...)

Generates the 2D cross section of the tokamak build

Note

Manipulates data in dd.build

ActorCXbuild
├─ rebuild_wall
├─ layers_aware_of_pf_coils
├─ divertor_hfs_size_fraction
├─ divertor_lfs_size_fraction
├─ n_points
└─ do_plot

HFSsizing

FUSE.ActorHFSsizing — Method

ActorHFSsizing(dd::IMAS.dd, act::ParametersAllActors; kw...)

Actor that resizes the High Field Side of the tokamak radial build. It changes the radial build of the center stack (plug, OH, and TF) accounting for stresses, superconductors critical currents, flux swing, and field requirements.

Note

Manipulates radial build information in dd.build.layer

ActorHFSsizing
├─ error_on_technology
├─ error_on_performance
├─ do_plot
└─ verbose

FluxSwing

FUSE.ActorFluxSwing — Method

ActorFluxSwing(dd::IMAS.dd, act::ParametersAllActors; kw...)

Depending on operate_oh_at_j_crit

true => Evaluate the OH current limits, and evaluate flattop duration from that.

false => Evaluate what are the currents needed for a given flattop duration. This may or may not exceed the OH current limits.

OH flux consumption based on:

rampup estimate based on Ejima coefficient

flattop consumption

vertical field from PF coils

Note

Stores data in dd.build.flux_swing, dd.build.tf, and dd.build.oh

ActorFluxSwing
└─ operate_oh_at_j_crit

LFSsizing

FUSE.ActorLFSsizing — Method

ActorLFSsizing(dd::IMAS.dd, act::ParametersAllActors; kw...)

Actor that resizes the Low Field Side of the tokamak radial build It changes the location of the outer TF leg by takig into account requirement of ripple and maintenance ports.

Note

Manipulates radial build information in dd.build.layer

ActorLFSsizing
├─ maintenance
├─ tor_modularity
├─ pol_modularity
├─ do_plot
└─ verbose

Stresses

FUSE.ActorStresses — Method

ActorStresses(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates mechanical stresses on the center stack

Note

Stores data in dd.solid_mechanics

ActorStresses
├─ do_plot
└─ n_points

Compound (3 actors)

DynamicPlasma

FUSE.ActorDynamicPlasma — Method

ActorDynamicPlasma(dd::IMAS.dd, act::ParametersAllActors; kw...)

Compound evolves plasma in time

ActorDynamicPlasma
├─ Δt  [s]
├─ Nt
├─ evolve_transport
├─ evolve_pedestal
├─ evolve_hcd
├─ evolve_current
├─ evolve_equilibrium
├─ evolve_pf_active
├─ ip_controller
├─ time_derivatives_sources
└─ verbose

StationaryPlasma

FUSE.ActorStationaryPlasma — Method

ActorStationaryPlasma(dd::IMAS.dd, act::ParametersAllActors; kw...)

Compound actor that runs the following actors in succesion to find a self-consistent stationary plasma solution

ActorCurrent

ActorHCD

ActorCoreTransport

ActorEquilibrium

Note

Stores data in dd.equilibrium, dd.core_profiles, dd.core_sources, dd.core_transport

ActorStationaryPlasma
├─ max_iterations
├─ convergence_error
├─ do_plot
└─ verbose

WholeFacility

FUSE.ActorWholeFacility — Method

ActorWholeFacility(dd::IMAS.dd, act::ParametersAllActors; kw...)

Compound actor that runs all the physics, engineering and costing actors needed to model the whole plant:

ActorStationaryPlasma

ActorPlasmaLimits

ActorHFSsizing

ActorLFSsizing

ActorCXbuild

ActorFluxSwing

ActorStresses

ActorPFdesign

ActorPFactive

ActorPassiveStructures

ActorVerticalStability

ActorNeutronics

ActorBlanket

ActorDivertors

ActorBalanceOfPlant

ActorCosting

Note

Stores data in dd

ActorWholeFacility
├─ update_plasma
└─ update_build

Control (1 actors)

ControllerIp

FUSE.ActorControllerIp — Method

ActorControllerIp(dd::IMAS.dd, act::ParametersAllActors; kw...)

Controls the loop voltage Vloop to obtain the target plasma cuurent Ip pulse_schedule

Algorithm names that have "PID" in their name will be considered linear and operate on dd.controllers.linear_controller otherwise dd.controllers.nonlinear_controller

Note

External controllers can be defined by dispatching on the controllers__linearcontroller functors (note, yes "PID" in the controlalgorithm)

function (controller::IMAS.controllers__linear_controller{T})(control_name::Val{:ip}, control_algorithm::Val{:my_PID}, setpoint::T, value::T, time0::Float64) where {T<:Real}
    ...
end

or controllers__nonlinearcontroller functors (note, no "PID" in the controlalgorithm)

function (controller::IMAS.controllers__nonlinear_controller{T})(control_name::Val{:ip}, control_algorithm::Val{:my_MPC}, setpoint::T, value::T, time0::Float64) where {T<:Real}
    ...
end

ActorControllerIp
├─ algorithm
├─ P
├─ I
├─ D
└─ Vloop_initial  [V]

Costing (3 actors)

Costing

FUSE.ActorCosting — Method

ActorCosting(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates the cost of building, operating, and recommission the fusion power plant.

Note

Stores data in dd.costing

ActorCosting
├─ model
├─ construction_start_year  [year]
├─ future_inflation_rate
├─ plant_lifetime  [year]
├─ availability
├─ production_increase
└─ learning_rate

CostingARIES

FUSE.ActorCostingARIES — Method

ActorCostingARIES(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates costing based on ARIES cost account documentation https://cer.ucsd.edu/_files/publications/UCSD-CER-13-01.pdf

Note

Stores data in dd.costing

ActorCostingARIES
├─ land_space  [acres]
├─ building_volume  [m³]
├─ interest_rate
├─ indirect_cost_rate
├─ escalation_fraction
└─ blanket_lifetime  [year]

CostingSheffield

FUSE.ActorCostingSheffield — Method

ActorCostingSheffield(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates costing based on Sheffield and Milora, FS&T 70 (2016)

Note

Stores data in dd.costing

ActorCostingSheffield
├─ construction_lead_time  [year]
├─ fixed_charge_rate
├─ capitalize_blanket
├─ capitalize_divertor
├─ divertor_fluence_lifetime  [MW*yr/m²]
└─ blanket_fluence_lifetime  [MW*yr/m²]

Current (3 actors)

Current

FUSE.ActorCurrent — Method

ActorCurrent(dd::IMAS.dd, act::ParametersAllActors; kw...)

Provides a common interface to run multiple ohmic current evolution actors

The fundamental quantitiy being solved is j_total in dd.core_profiles.profiles_1d[]

Also updates:

j_ohmic, j_tor in dd.core_profiles.profiles_1d[]

j_parallel in dd.equilibrium.time_slice[].profiles_1d

bootstrap and ohmic parallel current and heating sources in dd.core_sources

ActorCurrent
├─ model
├─ ip_from
└─ vloop_from

QED

FUSE.ActorQED — Method

ActorQED(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evolves the plasma current using the QED current diffusion solver.

The fundamental quantitiy being solved is j_total in dd.core_profiles.profiles_1d[]

Note

This actor operates at "dd.global_time", any time advance must be done outside of the actor

IMAS.new_timeslice!(dd, dd.global_time + Δt)
dd.global_time += Δt
ActorQED(dd, act)

ActorQED
├─ Δt  [s]
├─ Nt
├─ solve_for
├─ qmin_desired
├─ ip_from
└─ vloop_from

SteadyStateCurrent

FUSE.ActorSteadyStateCurrent — Method

ActorSteadyStateCurrent(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evolves the ohmic current to steady state using the conductivity from dd.core_profiles

Note

The fundamental quantitiy being solved is j_total in dd.core_profiles.profiles_1d[]

ActorSteadyStateCurrent
├─ allow_floating_plasma_current
├─ current_relaxation_radius
└─ ip_from

Divertors (1 actors)

Divertors

FUSE.ActorDivertors — Method

ActorDivertors(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evaluates divertor loading and deposited power

Note

Stores data in dd.divertors

ActorDivertors
├─ heat_flux_model
├─ impurities
├─ impurities_fraction
├─ heat_spread_factor
├─ thermal_power_extraction_efficiency
└─ verbose

Equilibrium (5 actors)

CHEASE

FUSE.ActorCHEASE — Method

ActorCHEASE(dd::IMAS.dd, act::ParametersAllActors; kw...)

Runs the Fixed boundary equilibrium solver CHEASE

ActorCHEASE
├─ free_boundary
├─ clear_workdir
└─ rescale_eq_to_ip

EGGO

FUSE.ActorEGGO — Method

ActorEGGO(dd::IMAS.dd, act::ParametersAllActors; kw...)

Runs the Fixed boundary equilibrium solver EGGO

ActorEGGO
├─ model
├─ do_plot
└─ debug

Equilibrium

FUSE.ActorEquilibrium — Method

ActorEquilibrium(dd::IMAS.dd, act::ParametersAllActors; kw...)

Provides a common interface to run different equilibrium actors

ActorEquilibrium
├─ model
├─ symmetrize
├─ j_p_from
├─ ip_from
├─ vacuum_r0_b0_from
└─ do_plot

FRESCO

FUSE.ActorFRESCO — Method

ActorFRESCO(dd::IMAS.dd, act::ParametersAllActors; kw...)

Runs the Fixed boundary equilibrium solver FRESCO

ActorFRESCO
├─ control
├─ number_of_iterations
├─ relax
├─ tolerance
├─ fixed_grid
├─ nR
├─ nZ
├─ do_plot
└─ debug

TEQUILA

FUSE.ActorTEQUILA — Method

ActorTEQUILA(dd::IMAS.dd, act::ParametersAllActors; kw...)

Runs the Fixed boundary equilibrium solver TEQUILA

ActorTEQUILA
├─ free_boundary
├─ number_of_radial_grid_points
├─ number_of_fourier_modes
├─ number_of_MXH_harmonics
├─ number_of_iterations
├─ relax
├─ tolerance
├─ fixed_grid
├─ R  [m]
├─ Z  [m]
├─ do_plot
└─ debug

Hcd (5 actors)

HCD

FUSE.ActorHCD — Method

ActorHCD(dd::IMAS.dd, act::ParametersAllActors; kw...)

Provides a common interface to run HCD actors

ActorHCD
├─ ec_model
├─ ic_model
├─ lh_model
├─ nb_model
├─ pellet_model
└─ neutral_model

NeutralFueling

FUSE.ActorNeutralFueling — Method

ActorNeutralFueling(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates the neutral fueling deposition

ActorNeutralFueling
├─ τp_over_τe
└─ T_wall  [eV]

SimpleIC

FUSE.ActorSimpleIC — Method

ActorSimpleIC(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates the ion-cyclotron electron/ion energy deposition and current drive as a gaussian.

NOTE: Current drive efficiency from GASC, based on "G. Tonon 'Current Drive Efficiency Requirements for an Attractive Steady-State Reactor'"

Note

Reads data in dd.ic_antennas, dd.pulse_schedule and stores data in dd.waves and dd.core_sources

ActorSimpleIC
└─ actuator

SimpleLH

FUSE.ActorSimpleLH — Method

ActorSimpleLH(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates the Lower-hybrid electron energy deposition and current drive as a gaussian.

NOTE: Current drive efficiency from GASC, based on "G. Tonon 'Current Drive Efficiency Requirements for an Attractive Steady-State Reactor'"

Note

Reads data in dd.lh_antennas, dd.pulse_schedule and stores data in dd.waves and dd.core_sources

ActorSimpleLH
└─ actuator

SimplePL

FUSE.ActorSimplePL — Method

ActorSimplePL(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates the Pellet particle deposition

Note

Reads data in dd.pellet_launchers, dd.pulse_schedule and stores data in dd.core_sources

ActorSimplePL
└─ actuator

Nuclear (2 actors)

Blanket

FUSE.ActorBlanket — Method

ActorBlanket(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evaluates blankets tritium breeding ratio (TBR), heat deposition, and neutron leakage

Note

Stores data in dd.blanket

ActorBlanket
├─ minimum_first_wall_thickness  [m]
├─ blanket_multiplier
├─ thermal_power_extraction_efficiency
├─ max_Li6_enrichment_fraction
└─ verbose

Neutronics

FUSE.ActorNeutronics — Method

ActorNeutronics(dd::IMAS.dd, act::ParametersAllActors; kw...)

Estimates the neutron wall loading

Note

Stores data in dd.neutronics

ActorNeutronics
├─ N
└─ do_plot

Pedestal (3 actors)

Pedestal

FUSE.ActorPedestal — Method

ActorPedestal(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evaluates the pedestal boundary condition (height and width)

ActorPedestal
├─ rho_nml
├─ rho_ped
├─ T_ratio_pedestal
├─ Te_sep
├─ ip_from
├─ βn_from
├─ ne_from
├─ zeff_from
├─ mode_transitions  [s]
├─ density_match
├─ model
├─ tau_t  [s]
├─ tau_n  [s]
├─ density_ratio_L_over_H
├─ zeff_ratio_L_over_H
└─ do_plot

EPED

FUSE.ActorEPED — Method

ActorEPED(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evaluates the pedestal boundary condition (height and width)

ActorEPED
├─ rho_nml
├─ rho_ped
├─ T_ratio_pedestal
├─ Te_sep
├─ ip_from
├─ βn_from
├─ ne_from
├─ zeff_from
├─ ped_factor
├─ only_powerlaw
└─ warn_nn_train_bounds

WPED

FUSE.ActorWPED — Method

ActorWPED(dd::IMAS.dd, act::ParametersAllActors; kw...)

Finds the temperature profile at the edge to match the pedtocorefraction of stored energy set in par.pedtocorefraction

ActorWPED
├─ rho_nml
├─ rho_ped
├─ T_ratio_pedestal
├─ Te_sep
├─ ip_from
├─ βn_from
├─ ne_from
├─ zeff_from
├─ ped_to_core_fraction
└─ do_plot

Pf (3 actors)

PFdesign

FUSE.ActorPFdesign — Method

ActorPFdesign(dd::IMAS.dd, act::ParametersAllActors; kw...)

Optimize PF coil locations to achieve desired equilibrium

Note

Manupulates data in dd.pf_active

ActorPFdesign
├─ symmetric
├─ update_equilibrium
├─ model
├─ reset_rails
├─ do_plot
└─ verbose

PFactive

FUSE.ActorPFactive — Method

ActorPFactive(dd::IMAS.dd, act::ParametersAllActors; kw...)

Finds the optimal coil currents to match the equilibrium boundary shape

Note

Manupulates data in dd.pf_active

ActorPFactive
├─ green_model
├─ update_equilibrium
├─ x_points_weight
├─ strike_points_weight
└─ do_plot

PassiveStructures

FUSE.ActorPassiveStructures — Method

ActorPassiveStructures(dd::IMAS.dd, act::ParametersAllActors; kw...)

Populates pf_passive IDS based on the vacuum vessel layer(s)

ActorPassiveStructures
├─ wall_precision
├─ min_n_segments
└─ do_plot

Stability (2 actors)

PlasmaLimits

FUSE.ActorPlasmaLimits — Method

ActorPlasmaLimits(dd::IMAS.dd, act::ParametersAllActors; kw...)

Runs all the limit actors

ActorPlasmaLimits
├─ models
├─ raise_on_breach
└─ verbose

VerticalStability

FUSE.ActorVerticalStability — Method

ActorVerticalStability(dd::IMAS.dd, act::ParametersAllActors; kw...)

Compute vertical stability metrics

ActorVerticalStability
├─ model
└─ do_plot

Transport (8 actors)

CoreTransport

FUSE.ActorCoreTransport — Method

ActorCoreTransport(dd::IMAS.dd, act::ParametersAllActors; kw...)

Provides a common interface to run multiple core transport actors

ActorCoreTransport
├─ model
└─ do_plot

FluxCalculator

FUSE.ActorFluxCalculator — Method

ActorFluxCalculator(dd::IMAS.dd, act::ParametersAllActors; kw...)

Provides a common interface to run multiple transport model actors

ActorFluxCalculator
├─ rho_transport
├─ turbulence_model
└─ neoclassical_model

FluxMatcher

FUSE.ActorFluxMatcher — Method

ActorFluxMatcher(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evalutes the transport fluxes and source fluxes and minimizes the flux_match error

ActorFluxMatcher
├─ rho_transport
├─ evolve_Ti
├─ evolve_Te
├─ evolve_densities
├─ evolve_rotation
├─ evolve_pedestal
├─ evolve_plasma_sources
├─ find_widths
├─ max_iterations
├─ algorithm
├─ step_size
├─ Δt  [s]
├─ relax
├─ scale_turbulence_law
├─ scale_turbulence_value
├─ do_plot
└─ verbose

AnalyticTurbulence

FUSE.ActorAnalyticTurbulence — Method

ActorAnalyticTurbulence(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evaluates analytic turbulence models

ActorAnalyticTurbulence
├─ model
└─ rho_transport

EPEDprofiles

FUSE.ActorEPEDprofiles — Method

ActorEPEDprofiles(dd::IMAS.dd, act::ParametersAllActors; kw...)

Updates pedestal height and width and blends with core profiles that are defined by shaping factors.

Does not change on-axis values of plasma profiles.

ActorEPEDprofiles
├─ Te_shaping
├─ ne_shaping
├─ T_ratio_pedestal
└─ T_ratio_core

Neoclassical

FUSE.ActorNeoclassical — Method

ActorNeoclassical(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evaluates the neoclassical transport fluxes

ActorNeoclassical
├─ model
└─ rho_transport

QLGYRO

FUSE.ActorQLGYRO — Method

ActorQLGYRO(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evaluates the QLGYRO predicted turbulence

ActorQLGYRO
├─ ky
├─ nky
├─ cpu_per_ky
├─ kygrid_model
├─ sat_rule
├─ n_field
├─ delta_t
├─ max_time
├─ rho_transport
└─ lump_ions

TGLF

FUSE.ActorTGLF — Method

ActorTGLF(dd::IMAS.dd, act::ParametersAllActors; kw...)

Evaluates the TGLF predicted turbulence

ActorTGLF
├─ model
├─ onnx_model
├─ sat_rule
├─ electromagnetic
├─ tglfnn_model
├─ rho_transport
├─ warn_nn_train_bounds
├─ custom_input_files
├─ lump_ions
├─ save_input_tglfs_to_folder
└─ debug

Wall loading (2 actors)

CoreRadHeatFlux

FUSE.ActorCoreRadHeatFlux — Method

ActorCoreRadHeatFlux(dd::IMAS.dd, act::ParametersAllActors; kw...)

Computes the heat flux on the wall due to the core radiation

ActorCoreRadHeatFlux
├─ N
├─ r  [m]
├─ q  [W m⁻²]
├─ levels
├─ merge_wall
├─ step  [m]
└─ do_plot

ParticleHeatFlux

FUSE.ActorParticleHeatFlux — Method

ActorParticleHeatFlux(dd::IMAS.dd, act::ParametersAllActors; kw...)

Computes the heat flux on the wall due to the charged particles

ActorParticleHeatFlux
├─ r  [m]
├─ q  [W m⁻²]
├─ levels
├─ merge_wall
├─ step  [m]
└─ do_plot