Use cases

FUSE comes with a set of pre-cooked use cases. The case_parameters(:use_case, ...) method returns the ini and act parameters for that specific use_case. These ini and act can then be further customized before running a FUSE simulation.

To create your own case and add them to FUSE/cases copy one of the other cases as a template and change the ini/act parameters inside. A handy way of generating the ini code from your_ini that you created in a notebook or elsewhere is to call the function FUSE.case_parameter_creation_from_ini(your_ini) which will return a nicely formatted code snippet.

Tip!

Click on the Source button of each use case to see how each is setup

FUSE.case_parameters — Function

case_parameters(:ARC; flux_matcher::Bool=false)

CFS/MIT ARC design

case_parameters(:CAT)

GA Compact Advanced Tokamak design

case_parameters(::Val{:D3D}, shot::Int;
    fit_profiles::Bool=true,
    EFIT_tree::String="EFIT02",
    PROFILES_tree::String="ZIPFIT01",
    CER_analysis_type::String="CERAUTO",
    EFIT_run_id::String="",
    PROFILES_run_id::String="",
    omfit_host::String=get(ENV, "FUSE_OMFIT_HOST", "somega.gat.com"),
    omfit_root::String=get(ENV, "FUSE_OMFIT_ROOT", "/fusion/projects/theory/fuse/d3d_data_fetching/OMFIT-source"),
    omas_root::String=get(ENV, "FUSE_OMAS_ROOT", "/fusion/projects/theory/fuse/d3d_data_fetching/omas"),
    time_averaging::Float64=0.05,
    rho_averaging::Float64=0.25,
    new_impurity_match_power_rad::Symbol=:none,
    use_local_cache::Bool=false,
    use_interferometer::Bool=true
)

DIII-D from experimental shot

case_parameters(::Val{:D3D}, ods_file::AbstractString)

DIII-D from ods file

case_parameters(::Val{:D3D}, dd::IMAS.dd)

DIII-D from dd file

case_parameters(::Val{:D3D}, scenario::Symbol)

DIII-D from sample cases

case_parameters(::Val{:D3D_machine})

Base DIII-D machine parameters that are then extended by the other case_parameters(:D3D, ...) functions

case_parameters(:DTT)

DTT

case_parameters(::Val{:EXCITE})

GA EXCITE design

case_parameters(:FIRST)

Formosa Integrated Research Spherical Tokamak

case_parameters(:FPP; flux_matcher::Bool=false)

GA's FPP design

case_parameters(::Val{:HDB5}; tokamak::Symbol=:all, case::Int=0, database_case::Int=0)

For description of cases/variables see https://osf.io/593q6/

case_parameters(
	:ITER,
	init_from::Symbol,
	boundary_from::Symbol=:auto,
	ne_setting::Symbol=:ne_ped,
	time_dependent::Bool=false)

ITER

case_parameters(:JET)

JET D-T case

case_parameters(:KDEMO)

case_parameters(:KSTAR; init_from::Symbol)

case_parameters(:MANTA; flux_matcher::Bool=false)

MANTA (Modular Adjustable Negative-Triangularity ARC)

https://arxiv.org/abs/2405.20243

https://burningplasma.org/resources/ref/Web_Seminars/MANTA_USBPO_Webinar_Presentation.pdf

case_parameters(:MASTU; init_from::Symbol)

case_parameters(:NSTX; init_from::Symbol)

case_parameters(:SPARC; flux_matcher::Bool=false)

CFS/MIT SPARC design

case_parameters(::Val{:STEP}; init_from::Symbol=:scalars, pf_from::Symbol=init_from)

UKAEA STEP design

case_parameters(::Val{:UNIT})

UNIT circular plasma

case_parameters(::Val{:baby_MANTA}; flux_matcher::Bool=false)

case_parameters(case::Symbol, args...; kw...)

return ini and act for a use-case

NOTE: if case starts with test__ then the regression test cases are loaded