Use cases

FUSE comes with a set of pre-cookes used cases. The case_parameters(:use_case, ...) method returns the ini and act parameters for that specific use_case. These ini and act can then be further customized before running a FUSE simulation.

To create your own case and add them to FUSE/cases copy one of the other cases as a template and change the ini/act parameters inside. A handy way of generating the ini code from your_ini that you created in a notebook or elsewhere is to call the function FUSE.case_parameter_creation_from_ini(your_ini) which will return a nicely formatted code snippet.

Tip!

Click on the Source button of each use case to see how each is setup

ARC

FUSE.case_parameters — Method

case_parameters(:ARC; flux_matcher::Bool=false)

CFS/MIT ARC design

CAT

FUSE.case_parameters — Method

case_parameters(:CAT)

GA Compact Advanced Tokamak design

D3D

FUSE.case_parameters — Method

case_parameters(:D3D;
	scenario::Union{Symbol,AbstractString}=:default,
	scenario_sources::Bool=true,
	scenario_core_profiles::Bool=true)

DIII-D

Scenario keyword can be:

:default 133221

:H_mode a prototypical H_mode

:L_mode a prototypical L_mode

a user defined string pointing to a ODS on file saved in JSON format

scenario_sources keywods says whether core_sources will be taken from scenario or not

DTT

FUSE.case_parameters — Method

case_parameters(:DTT)

DTT

EXCITE

FUSE.case_parameters — Method

case_parameters(::Type{Val{:EXCITE}})::Tuple{ParametersAllInits,ParametersAllActors}

GA EXCITE design

FIRST

FUSE.case_parameters — Method

case_parameters(:FIRST)

Formosa Integrated Research Spherical Tokamak

FPP

FUSE.case_parameters — Method

case_parameters(:FPP; flux_matcher::Bool=false)

GA's FPP design

HDB5

FUSE.case_parameters — Method

case_parameters(::Type{Val{:HDB5}}; tokamak::Union{String,Symbol}=:any, case=missing, database_case=missing)

For description of cases/variables see https://osf.io/593q6/

ITER

FUSE.case_parameters — Method

case_parameters(
	:ITER,
	init_from::Symbol,
	boundary_from::Symbol=:auto,
	ne_setting::Symbol=:ne_ped,
	time_dependent::Bool=false)

ITER

KDEMO

FUSE.case_parameters — Method

case_parameters(:KDEMO)

KDEMO_compact

Missing docstring.

Missing docstring for case_parameters(::Type{Val{:KDEMO_compact}}). Check Documenter's build log for details.

MANTA

FUSE.case_parameters — Method

case_parameters(:MANTA; flux_matcher::Bool=false)

MANTA (Modular Adjustable Negative-Triangularity ARC)

https://arxiv.org/abs/2405.20243

https://burningplasma.org/resources/ref/WebSeminars/MANTAUSBPOWebinarPresentation.pdf

SPARC

FUSE.case_parameters — Method

case_parameters(:SPARC; flux_matcher::Bool=false)

CFS/MIT SPARC design

STEP

FUSE.case_parameters — Method

case_parameters(::Type{Val{:STEP}}; init_from::Symbol=:scalars, pf_from::Symbol=init_from)

UKAEA STEP design

baby_MANTA

FUSE.case_parameters — Method

case_parameters(::Type{Val{:baby_MANTA}}; flux_matcher::Bool=false)